+++ source/XYO/Fabricare/Internal/platform/msys2-mingw32.js
@@ -1,9 +1,9 @@
 // Created by Grigore Stefan <g_stefan@yahoo.com>
 // Public domain (Unlicense) <http://unlicense.org>
-// SPDX-FileCopyrightText: 2021-2024 Grigore Stefan <g_stefan@yahoo.com>
+// SPDX-FileCopyrightText: 2021-2025 Grigore Stefan <g_stefan@yahoo.com>
 // SPDX-License-Identifier: Unlicense
 
-Platform.name = "sys-mingw32";
+Platform.name = "msys2-mingw32";
 Platform.next = "mingw32";
 
-Fabricare.include("platform/sys-mingw.run");
+Fabricare.include("platform/msys2-mingw.run");