+++ source/XYO/Fabricare/Internal/platform/msys2-ucrt64.js
@@ -3,13 +3,7 @@
 // SPDX-FileCopyrightText: 2021-2025 Grigore Stefan <g_stefan@yahoo.com>
 // SPDX-License-Identifier: Unlicense
 
-Fabricare.include("solution/xyo-cpp.library");
+Platform.name = "msys2-ucrt64";
+Platform.next = "ucrt64";
 
-prepareProjects();
-selectMainProject();
-
-if (Fabricare.include(Fabricare.action)) {
-	return;
-};
-
-exit(1, "Action " + Fabricare.action + " not found!");
+Fabricare.include("platform/msys2-mingw.run");