+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [7.4.0] - 2024-12-04
+
+### Added
+
+- platform msys2-ucrt64
+- platform ucrt64
+
+### Changed
+
+- renamed platform name from sys-mingw32 to msys2-mingw32
+- renamed platform name from sys-mingw64 to msys2-mingw64
+
 ## [7.0.0] - 2024-08-20
 
 ### Removed